Refactor traits to break out into separate files as the main feature context was starting to get big.
